fix(HappyHare): Retains the correct font size for filament status messages (#2428)
* Moved filament status message inside of SVG component for consistent font sizing

* updates from linter and prettifier

* moved swap count lookup getters

* Ensure capitalize method is imported
