+++ src/components/panels/MmuPanel.vue
@@ -82,8 +82,7 @@
             </v-row>
             <v-row>
                 <v-col :cols="col1Size">
-                    <div class="text--disabled body-2">{{ toolchangeText }}</div>
-                    <div class="text-center body-1">{{ statusText }}</div>
+                    <div class="text--disabled body-1">{{ toolchangeText }}</div>
                     <mmu-filament-status />
                     <div v-if="showClogDetection" class="text-center">
                         <mmu-clog-meter v-if="hasMmuEncoder" width="40%" />
@@ -135,14 +134,7 @@
 <script lang="ts">
 import { Component, Mixins } from 'vue-property-decorator'
 import BaseMixin from '@/components/mixins/base'
-import MmuMixin, {
-    ACTION_IDLE,
-    ACTION_LOADING,
-    ACTION_UNLOADING,
-    TOOL_GATE_BYPASS,
-    TOOL_GATE_UNKNOWN,
-} from '@/components/mixins/mmu'
-import { capitalize } from '@/plugins/helpers'
+import MmuMixin, { TOOL_GATE_BYPASS, TOOL_GATE_UNKNOWN } from '@/components/mixins/mmu'
 import { mdiMulticast, mdiDotsVertical, mdiCheckAll, mdiNoteText, mdiInformationOutline, mdiRefresh } from '@mdi/js'
 import Panel from '@/components/ui/Panel.vue'
 import MmuPanelSettings from '@/components/panels/Mmu/MmuPanelSettings.vue'
@@ -226,18 +218,6 @@ export default class MmuPanel extends Mixins(BaseMixin, MmuMixin) {
         return this.$store.state.gui.view.mmu.showDetails ?? true
     }
 
-    get slicerToolMap() {
-        return this.mmu?.slicer_tool_map ?? undefined
-    }
-
-    get totalToolchanges() {
-        return this.slicerToolMap?.total_toolchanges ?? 0
-    }
-
-    get numToolchanges() {
-        return this.mmu?.num_toolchanges ?? 0
-    }
-
     get toolchangeText() {
         if (this.nextTool === TOOL_GATE_UNKNOWN) return ''
 
@@ -258,35 +238,8 @@ export default class MmuPanel extends Mixins(BaseMixin, MmuMixin) {
         return this.mmu?.next_tool ?? TOOL_GATE_UNKNOWN
     }
 
-    get statusText() {
-        if (['complete', 'error', 'cancelled', 'started'].includes(this.mmuPrintState)) {
-            return capitalize(this.mmuPrintState)
-        }
-
-        if ([ACTION_LOADING, ACTION_UNLOADING].includes(this.mmuAction)) {
-            return `${this.mmuAction}: ${this.filamentPosition}mm`
-        }
-
-        if (this.mmuAction !== ACTION_IDLE) return this.mmuAction
-
-        if (this.mmuPrintState === 'printing') {
-            let str = `Printing (${this.numToolchanges}`
-            if (this.totalToolchanges > 0) str += `/${this.totalToolchanges}`
-            str += ' swaps)'
-            return str
-        }
-
-        const filament = this.mmu?.filament ?? 'Unknown'
-
-        return filament !== 'Unloaded' ? `Filament: ${this.filamentPosition}mm` : 'Filament: Unloaded'
-    }
-
     get reasonForPause() {
         return this.mmu?.reason_for_pause ?? null
-    }
-
-    get filamentPosition() {
-        return (this.mmu?.filament_position ?? 0).toFixed(1)
     }
 
     get fileForTtgMap() {
